Services/AzureFunctionWAVConversion/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob==12.8.0
cryptography==2.8
ffmpeg-python==0.2.0